Implement an HvpSettings class
Moved the SQLite code inside the new class HvpDataConnector in order to
get access to the same database to store settings inside it.

The class HvpSettings only does a read/write in a "settings" table.

Updated the code to use the new class.

The Configuration is now editable (title and video-backend).